gateway-service/requirements.txt
# Gateway Service Dependencies
# 基于原ZZDSJ Backend API项目的依赖进行筛选和优化

# FastAPI核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证和配置
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# JWT认证
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# 缓存和数据库连接
redis==5.0.1
aioredis==2.0.1

# 日志和监控
structlog==23.2.0
loguru==0.7.2

# 异步任务和并发
asyncio-pool==0.6.0
asyncio-throttle==1.0.2

# 工具库
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3

# 开发工具
python-dotenv==1.0.0

# 网络和安全
certifi==2023.11.17
urllib3==2.1.0

# 数据处理
orjson==3.9.10

# 监控和健康检查
prometheus-client==0.19.0

# 配置管理
pyyaml==6.0.1
toml==0.10.2
